Limit interface zone condition to a single interface
The nftables rules to match interface zones do a bit mask comparison
of the mark and the supplied condition. These constructs don't
support testing a list of interfaces so I changed multiSelect to
false for these condition definitions. I also restored the natEgress
flag for VLAN's to the previous default value until we have a chance
to re-think our VLAN stuff.