Add paste/import wireguard configuration logic
I added an Import icon to the Wireguard interface creation dialog
that allows pasting JSON wireguard configs from NGFW. While we'll
likely go with a paste/detection model similar to how this is
handled in NGFW I figured some of this may still be useful. This
seems to be the only instance of using Ext.Msg.prompt which could
prove useful in other cases.